refactor: refactor de toute la section campus
