+++ src/features/Campus/Crous/CrousMenuScreen.tsx
@@ -3,7 +3,7 @@ import { View, Text, ScrollView, ActivityIndicator, TouchableOpacity, FlatList }
 import { SafeAreaView, useSafeAreaInsets } from 'react-native-safe-area-context';
 import { MaterialCommunityIcons } from '@expo/vector-icons';
 
-import { CrousService, CrousDayMenu } from './CrousService';
+import { CrousService, CrousDayMenu } from '../services/CrousService';
 import style, { tokens } from '../../../shared/theme/Theme';
 import { AppContext } from '../../../shared/services/AppCore';
 import Translator from '../../../shared/i18n/Translator';
